% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Divide.R
\name{Divide}
\alias{Divide}
\title{Divide() Function}
\usage{
Divide(x, y)
}
\arguments{
\item{a}{,b : two numbers to be operated}
}
\description{
This function divide a by b
}
\examples{
Divide()
}
\keyword{Divide}
