% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiply.R
\name{Multiply}
\alias{Multiply}
\title{Multiply() Function}
\usage{
Multiply(x, y)
}
\arguments{
\item{a}{,b : two numbers to be operated}
}
\description{
This function multiplies a and b
}
\examples{
Multiply()
}
\keyword{Add}
