% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subtract.R
\name{Subtract}
\alias{Subtract}
\title{Subtract Function}
\usage{
Subtract(x, y)
}
\arguments{
\item{a}{,b : two numbers to be operated}
}
\description{
This function subtrac a from b
}
\examples{
Subtract(3, 2)
}
\keyword{a}
\keyword{b}
\keyword{from}
\keyword{subtract}
